adding image URL to monster interface, and loading up this with a link to the assets of that monsters name.png
then adding this as a background image to the monster on initiative timeline

additionally adding a combat log service, to just hold a store of all the attacks generated

finally, moved the observable stuff into an abstract class, that the encounter handler inherits. this lets me take the boilerplate code out of the way. i tried to have an observable combat log too, but you can only inherit once in angular.